// packages/cooder-themes/type.ts
export type HighlightPalette = {
  mint: string;
  turquoise: string;
  bluejay: string;
  cerulean: string;
  radiantYellow: string;
  sunshine: string;
  blueberry: string;
  candy: string;
};

export type Shades = {
  '4p': string;
  '8p': string;
  '12p': string;
  '30p': string;
  '50p': string;
};

export type ShadesPalette = {
  primary: Shades;
  secondary: Shades;
};